Get consistent on _ and use __init__ for shared modules
