+++ sdk/mixedreality/azure-mixedreality-authentication/azure/mixedreality/authentication/_shared/aio/__init__.py
@@ -1,0 +1,5 @@
+# -------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for
+# license information.
+# --------------------------------------------------------------------------
