+++ sdk/mixedreality/azure-mixedreality-authentication/azure/mixedreality/authentication/_shared/aio/mixed_reality_token_credential.py
@@ -6,15 +6,19 @@
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    # pylint: disable=unused-import,ungrouped-imports
-    from typing import Any, Union
-    from azure.core.credentials import AccessToken, TokenCredential
+    from typing import Any
+    from azure.core.credentials import AccessToken
+    from azure.core.credentials_async import AsyncTokenCredential
 
-from ._static_access_token_credential import StaticAccessTokenCredential
-from .._client import MixedRealityStsClient
+from .static_access_token_credential import StaticAccessTokenCredential
+from ...aio._client_async import MixedRealityStsClient
 
-def get_mixedreality_credential(account_id, account_domain, endpoint_url, credential, **kwargs):
-        # type: (str, str, str, TokenCredential, Any) -> TokenCredential
+def get_mixedreality_credential(
+    account_id: str,
+    account_domain: str,
+    endpoint_url: str,
+    credential: "AsyncTokenCredential",
+    **kwargs):
         if isinstance(credential, StaticAccessTokenCredential):
             return credential
 
@@ -35,8 +39,7 @@ class MixedRealityTokenCredential(object):
     :param TokenCredential credential: The credential used to access the Mixed Reality service.
     """
 
-    def __init__(self, account_id, account_domain, endpoint_url, credential, **kwargs):
-        # type: (str, str, str, TokenCredential, Any) -> None
+    def __init__(self, account_id: str, account_domain: str, endpoint_url: str, credential: "AsyncTokenCredential", **kwargs):
         self.stsClient = MixedRealityStsClient(
             account_id=account_id,
             account_domain=account_domain,
@@ -44,6 +47,15 @@ class MixedRealityTokenCredential(object):
             credential=credential,
             **kwargs)
 
-    def get_token(self, *scopes, **kwargs):
-        # type: (*str, **Any) -> AccessToken
-        return self.stsClient.get_token(**kwargs)
+    async def get_token(self, *scopes: str, **kwargs: "Any") -> "AccessToken":
+        return await self.stsClient.get_token(**kwargs)
+
+    async def close(self) -> None:
+        self.stsClient.close()
+
+    async def __aenter__(self):
+        await self.stsClient.__aenter__()
+        return self
+
+    async def __aexit__(self, exc_type, exc_value, traceback) -> None:
+        await self.stsClient.__aexit__()
