+++ sdk/mixedreality/azure-mixedreality-authentication/azure/mixedreality/authentication/_utils/utils.py
@@ -6,10 +6,10 @@
 
 from azure.core.credentials import AccessToken
 
-from ._jwt import _retrieve_jwt_expiration_timestamp
+from .jwt import retrieve_jwt_expiration_timestamp
 from .._generated.models import StsTokenResponseMessage
 
-def _convert_to_access_token(token_response_message):
+def convert_to_access_token(token_response_message):
     # type: (StsTokenResponseMessage) -> AccessToken
     """
     Converts the specified token response message to an AccessToken.
@@ -17,6 +17,6 @@ def _convert_to_access_token(token_response_message):
     if not StsTokenResponseMessage:
         raise ValueError("token_response_message can not be None")
 
-    expiration_timestamp = _retrieve_jwt_expiration_timestamp(token_response_message.access_token)
+    expiration_timestamp = retrieve_jwt_expiration_timestamp(token_response_message.access_token)
 
     return AccessToken(token_response_message.access_token, expiration_timestamp)
