+++ sdk/mixedreality/azure-mixedreality-authentication/azure/mixedreality/authentication/aio/_client_async.py
@@ -20,9 +20,9 @@ from azure.core.pipeline.policies import AsyncBearerTokenCredentialPolicy
 from .._generated.aio import MixedRealityStsRestClient
 from .._generated.models import TokenRequestOptions
 from .._version import SDK_MONIKER
-from ..shared._authentication_endpoint import construct_endpoint_url
-from ..shared.aio._mixedreality_account_key_credential import MixedRealityAccountKeyCredential
-from ..utils import _convert_to_access_token, _generate_cv_base
+from .._shared.authentication_endpoint import construct_endpoint_url
+from .._shared.aio.mixedreality_account_key_credential import MixedRealityAccountKeyCredential
+from .._utils import convert_to_access_token, generate_cv_base
 
 if TYPE_CHECKING:
     from azure.core.credentials import AccessToken
@@ -94,14 +94,14 @@ class MixedRealityStsClient(object):
         :rtype: :class:`azure.core.credentials.AccessToken`
         """
         token_request_options = TokenRequestOptions()
-        token_request_options.client_request_id=_generate_cv_base()
+        token_request_options.client_request_id=generate_cv_base()
 
         response = await self._client.get_token(
             self._account_id,
             api_version=None,
             token_request_options=token_request_options,
             **kwargs)
-        return _convert_to_access_token(response)
+        return convert_to_access_token(response)
 
     async def close(self) -> None:
         await self._client.close()
