+++ sdk/mixedreality/azure-mixedreality-authentication/tests/test_cv.py
@@ -4,18 +4,18 @@
 # license information.
 # --------------------------------------------------------------------------
 
-from azure.mixedreality.authentication.utils._cv import _generate_cv_base
+from azure.mixedreality.authentication._utils.cv import generate_cv_base
 
 class TestCv:
     def test_generate_cv_base(self):
-        cv = _generate_cv_base()
+        cv = generate_cv_base()
 
         assert cv is not None
         assert len(cv) == 22
 
     def test_generate_cv_base_are_random(self):
-        cv1 = _generate_cv_base()
-        cv2 = _generate_cv_base()
+        cv1 = generate_cv_base()
+        cv2 = generate_cv_base()
 
         assert cv1 is not None
         assert cv2 is not None
